Rename some of the functions in the basemodel
This renames two of the functions in the Model class and updates them to
take as parameters a commit argument that determines whether the
session will actually be committed immediately following that function.

We now directoly call methods on the query attribute for the models,
allowing us to use get_or_404 and remove a few TODOs from the code

A few fixes were mode to the add_user functionality so that should
actually work (was not intending to do this, noticed some things while
working on it).