Update commands for copying the key file
The public key file should get coppied into the docker image, this
updates the commands so that will happen.

We first create a temporary file in the host machine with the contents
for the authorized_keys file.  After that, the file is copied into the
docker instance.  After that, we set the correct permissions (and
owner).  Finally, we clean up the temponary file.